Extract feature info support script into resource file
